- replaced all occurences of GPIO_PIN_NAMES by the newer const struct GPIO_SIGNAL_MAP and thus obsoleted the former one - implemented selection of trigger signal in parameterform (before it was only showing the current state)

+++ muondetector-gui/src/mainwindow.h
@@ -108,7 +108,7 @@ private slots:
 
     void onIpButtonClicked();
     void connected();
-    void sendInputSwitch(int id);
+    void sendInputSwitch(uint8_t id);
     void on_discr1Edit_editingFinished();
     void on_discr1Slider_sliderReleased();
     void on_discr1Slider_valueChanged(int value);
